+++ frontend/src/components/settings/backup/ImportExportModals.tsx
@@ -1,0 +1,176 @@
+import { Modal, Table, Tag as AntTag } from 'antd';
+
+export interface BackupDataYaml {
+  version: string;
+  created_at: string;
+  tags: { name: string; color: string }[];
+  todos: {
+    title: string;
+    prompt: string;
+    status: string;
+    executor?: string;
+    scheduler_enabled: boolean;
+    scheduler_config?: string;
+    tag_names: string[];
+    workspace?: string;
+  }[];
+}
+
+export interface ImportItem {
+  key: number;
+  title: string;
+  prompt: string;
+  status: string;
+  executor?: string;
+  scheduler_enabled: boolean;
+  scheduler_config?: string;
+  tag_names: string[];
+  workspace?: string;
+  action: 'new' | 'overwrite';
+  existingTitle?: string;
+}
+
+export function ImportExportModals({
+  wizardOpen, setWizardOpen, handleWizardConfirm, importing,
+  selectedRowKeys, setSelectedRowKeys, wizardItems,
+  exportModalOpen, setExportModalOpen, handleExportSelected,
+  exportingSelected, exportTodoKeys, setExportTodoKeys, todos,
+}: {
+  wizardOpen: boolean;
+  setWizardOpen: (v: boolean) => void;
+  handleWizardConfirm: () => Promise<void>;
+  importing: boolean;
+  selectedRowKeys: number[];
+  setSelectedRowKeys: (keys: number[]) => void;
+  wizardItems: ImportItem[];
+  exportModalOpen: boolean;
+  setExportModalOpen: (v: boolean) => void;
+  handleExportSelected: () => Promise<void>;
+  exportingSelected: boolean;
+  exportTodoKeys: number[];
+  setExportTodoKeys: (keys: number[]) => void;
+  todos: readonly any[];
+}) {
+  return (
+    <>
+      <Modal
+        title="导入预览"
+        open={wizardOpen}
+        onCancel={() => setWizardOpen(false)}
+        onOk={handleWizardConfirm}
+        okText={`导入 ${selectedRowKeys.length} 项`}
+        cancelText="取消"
+        confirmLoading={importing}
+        width={800}
+        okButtonProps={{ disabled: selectedRowKeys.length === 0 }}
+      >
+        <div style={{ marginBottom: 12, display: 'flex', gap: 16 }}>
+          <AntTag color="green">{wizardItems.filter(i => i.action === 'new').length} 个新建</AntTag>
+          <AntTag color="orange">{wizardItems.filter(i => i.action === 'overwrite').length} 个覆盖</AntTag>
+          <AntTag color="blue">已选 {selectedRowKeys.length} 项</AntTag>
+        </div>
+        <Table
+          dataSource={wizardItems}
+          rowKey="key"
+          size="small"
+          pagination={false}
+          scroll={{ y: 400 }}
+          rowSelection={{
+            selectedRowKeys,
+            onChange: (keys) => setSelectedRowKeys(keys as number[]),
+          }}
+          columns={[
+            {
+              title: '标题',
+              dataIndex: 'title',
+              ellipsis: true,
+              width: '35%',
+            },
+            {
+              title: '状态',
+              dataIndex: 'action',
+              width: 80,
+              render: (action: 'new' | 'overwrite') => (
+                <AntTag color={action === 'new' ? 'green' : 'orange'}>
+                  {action === 'new' ? '新建' : '覆盖'}
+                </AntTag>
+              ),
+            },
+            {
+              title: '执行器',
+              dataIndex: 'executor',
+              width: 100,
+              render: (v: string | undefined) => v || '-',
+            },
+            {
+              title: '标签',
+              dataIndex: 'tag_names',
+              width: 150,
+              render: (names: string[]) => names.length > 0
+                ? names.slice(0, 3).map(n => <AntTag key={n}>{n}</AntTag>)
+                : '-',
+            },
+            {
+              title: 'Prompt 摘要',
+              dataIndex: 'prompt',
+              ellipsis: true,
+              render: (v: string) => v ? v.slice(0, 60) + (v.length > 60 ? '...' : '') : '-',
+            },
+          ]}
+        />
+      </Modal>
+
+      <Modal
+        title="选择性导出"
+        open={exportModalOpen}
+        onCancel={() => setExportModalOpen(false)}
+        onOk={handleExportSelected}
+        okText={`导出 ${exportTodoKeys.length} 项`}
+        cancelText="取消"
+        confirmLoading={exportingSelected}
+        width={700}
+        okButtonProps={{ disabled: exportTodoKeys.length === 0 }}
+      >
+        <Table
+          dataSource={todos}
+          rowKey="id"
+          size="small"
+          pagination={{ pageSize: 50 }}
+          scroll={{ y: 400 }}
+          rowSelection={{
+            selectedRowKeys: exportTodoKeys,
+            onChange: (keys) => setExportTodoKeys(keys as number[]),
+          }}
+          columns={[
+            {
+              title: '标题',
+              dataIndex: 'title',
+              ellipsis: true,
+            },
+            {
+              title: '执行器',
+              dataIndex: 'executor',
+              width: 100,
+              render: (v: string | undefined) => v || '-',
+            },
+            {
+              title: '状态',
+              dataIndex: 'status',
+              width: 80,
+              render: (v: string) => {
+                const map: Record<string, { color: string; label: string }> = {
+                  pending: { color: 'default', label: '待办' },
+                  running: { color: 'processing', label: '进行中' },
+                  completed: { color: 'success', label: '完成' },
+                  failed: { color: 'error', label: '失败' },
+                };
+                const s = map[v] || { color: 'default', label: v };
+                return <AntTag color={s.color}>{s.label}</AntTag>;
+              },
+            },
+          ]}
+        />
+      </Modal>
+    </>
+  );
+}
